requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
db-dtypes==1.2.0
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.5
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-generativeai==0.7.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gotrue==2.5.4
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==2.0.0
openai==1.35.4
packaging==24.1
pandas==2.2.2
plotly==5.22.0
postgrest==0.16.8
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.6
redis==5.0.7
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
tenacity==8.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
zipp==3.19.2
